requirements.txt
pandas 
numpy
google-cloud-storage
PyYAML
seaborn
matplotlib
scikit-learn==1.3.2 
imbalanced-learn==0.11.0
ipykernel
statsmodels
xgboost 
lightgbm